chore: move components to correct folders
