--- client/src/components/CaseFooterSection/index.js
@@ -1,1 +1,0 @@
-export { CaseFooterSection } from './CaseFooterSection'
