--- client/src/components/Sidebar/index.js
@@ -1,1 +1,0 @@
-export { CardSidebar, Block, TitleSidebar, TitlePost, Date } from './Sidebar'
